Death, Rotation Limit, Sharks
There is now a limit on rotational speed, but it's slightly buggy or
jittery. (Might need a lerp somewhere). We can always change that back.

Also Sharks now spawn as the hunger number gets above 70. A red vignette
also starts to fade in. But there needs to be 1 sharkTimer total. Not 1
on each ring.